refactor(engine): move interfaces to API package
This commit moves two interfaces to the API package. The API package is meant to contain the public API of the engine. Both of these interfaces are implemented outside the engine.

Other interfaces may also need to be moved there, but the current focus is on ReadonlyStreamProcessorContext, StreamProcessorLifecycleAware was also moved, because it depends on ReadonlyStreamProcessorContext and is implemented in the broker
