+++ engine/src/main/java/io/camunda/zeebe/engine/api/StreamProcessorLifecycleAware.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.processing.streamprocessor;
+package io.camunda.zeebe.engine.api;
 
 public interface StreamProcessorLifecycleAware {
