+++ engine/src/main/java/io/camunda/zeebe/engine/state/immutable/ZeebeState.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.engine.state.immutable;
 
-import io.camunda.zeebe.engine.processing.streamprocessor.StreamProcessorLifecycleAware;
+import io.camunda.zeebe.engine.api.StreamProcessorLifecycleAware;
 import io.camunda.zeebe.engine.state.ZbColumnFamilies;
 
 public interface ZeebeState extends StreamProcessorLifecycleAware {
