+++ engine/src/test/java/io/camunda/zeebe/engine/state/migration/DbMigrationControllerTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
-import io.camunda.zeebe.engine.processing.streamprocessor.ReadonlyStreamProcessorContext;
+import io.camunda.zeebe.engine.api.ReadonlyStreamProcessorContext;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.BeforeEach;
